# apps/api/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: LibFind API
  version: 0.1.0-m0
  description: |
    Phase 0 OpenAPI foundation for LibFind.
    Base path: /api/v1
    Product marketplace endpoints remain Phase 1+.
servers:
  - url: /api/v1
    description: Versioned API
paths:
...
  /auth/otp/request:
    post:
      operationId: requestOtp
      summary: Create OTP challenge (M0 skeleton; SMS Phase 1)
      tags: [Auth]
      responses:
        '201':
          description: Challenge created
        '429':
          description: Rate limited
...
  /media/sign-upload:
    post:
      operationId: signUpload
      summary: Signed Cloudinary upload params (server-side secret)
      tags: [Media]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Signature payload without api_secret
        '503':
          description: Cloudinary not configured
  /spike/ssr-search:
    get:
      operationId: ssrSearchSpike
      summary: M0 SSR/search HTML spike (not product search)
      tags: [Spike]
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: HTML document
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token from OTP verify / refresh.